Fix coverage code.
