+++ cli/text_encoding.rs
@@ -54,6 +54,19 @@ pub fn strip_bom(text: &str) -> &str {
   }
 }
 
+pub fn source_map_from_code(code: &str) -> Option<Vec<u8>> {
+  static PREFIX: &str = "//# sourceMappingURL=data:application/json;base64,";
+  let last_line = code.rsplit(|u| u == '\n').next().unwrap();
+  if last_line.starts_with(PREFIX) {
+    let input = last_line.split_at(PREFIX.len()).1;
+    let decoded_map = base64::decode(input)
+      .expect("Unable to decode source map from emitted file.");
+    Some(decoded_map)
+  } else {
+    None
+  }
+}
+
 #[cfg(test)]
 mod tests {
   use super::*;
